add new module(Level) by Edison

+++ trunk/yzhd_bbs/src/com/ntsky/bbs/web/webwork/action/level/LevelsAction.java
@@ -1,5 +1,6 @@
 package com.ntsky.bbs.web.webwork.action.level;
 
+import com.ntsky.bbs.util.memory.LevelSingleton;
 import com.opensymphony.xwork.Preparable;
 
 public class LevelsAction extends LevelActionSupport implements Preparable {
@@ -10,18 +11,25 @@ public class LevelsAction extends LevelActionSupport implements Preparable {
 	
 	public String execute() throws Exception {
 		
-		/*if(!isPermisson("2_2")){
-			setWarnMessage("您没有论坛管理的权限.");
+		if(!isPermisson("4_4")){
+			setWarnMessage("您没有等级设置的权限.");
 			return NO_PERMISSION;
-		}*/
+		}
+		recordActLog("等级设置");
 		//this.setLevels(levelService.findLevels());
 		return SUCCESS;
     }
 	
 	public String doDelete() throws Exception {
+		if(!isPermisson("4_4")){
+			setWarnMessage("您没有等级设置的权限.");
+			return NO_PERMISSION;
+		}
+		recordActLog("等级设置");
 		//this.getParameters(id);
 		String id=this.getParameter("id");
 		levelService.removeLevel(levelService.getLevel(id));
+		LevelSingleton.getInstance().setLevels(levelService.findLevels());
 		return DELETE;
 	}
